requirements.txt
ttkbootstrap>=1.10.1
paho-mqtt==1.6.1